Intermediate release
Regression tests: a few bugs corrected (patch_input_code and anti-pgm overlay).

git-svn-id: http://cadia.ru.is/svn/repos/replicode@263 36b273a1-5dce-4f9d-8980-b90fca7d7570

+++ r_exec/view.inline.cpp
@@ -56,7 +56,7 @@ namespace	r_exec{
 	inline	View::View(bool		sync,
 						uint64	ijt,
 						float32	sln,
-						int16	res,
+						int32	res,
 						Code	*destination,
 						Code	*origin,
 						Code	*object):r_code::View(),controller(NULL){
@@ -68,7 +68,7 @@ namespace	r_exec{
 	inline	View::View(bool		sync,
 						uint64	ijt,
 						float32	sln,
-						int16	res,
+						int32	res,
 						Code	*destination,
 						Code	*origin,
 						Code	*object,
@@ -82,7 +82,7 @@ namespace	r_exec{
 	inline	void	View::init(bool		sync,
 								uint64	ijt,
 								float32	sln,
-								int16	res,
+								int32	res,
 								Code	*destination,
 								Code	*origin,
 								Code	*object){
